trax-analyzer/requirements.txt
pymupdf
pandas
openai
python-dotenv